refactor: Mutualize code for sending database and perf logs to support
Performance and Database email reporters were using the same duplicated
code

We want to use the same logic in the Logs email reporter, so it is time
to refactor this part to mutualize the code
